Allow to add gfp argument to blkdev_issue_flush() for old kernels
The memory allocation flag was removed from the arguments of
blkdev_issue_flush() in kernel 5.11.

This change was backported to RHEL 8.6 and later kernels.

To support older RHEL 8 kernel versions less than 8.6, this toggles
whether to pass a gfp argument to blkdev_issue_flush() depending on
the RHEL 8 minor version.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>