Add dummy error sector argument to blkdev_issue_flush() for old kernels
The blkdev_issue_flush function used to take an optional argument to
store the error location, but that was removed in kernel 5.7 since
it's not used by any callers.

This change was backported to RHEL 8.4 and later kernels.

To support older RHEL 8 kernel versions less than 8.4, this toggles
whether to add the error sector argument to blkdev_issue_flush()
depending on the RHEL 8 minor version.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>