Avoid using unexported __inode_attach_wb() in older kernels
inode_attach_wb() is not available from modules before kernel 5.3
since its entity function __inode_attach_wb() was not exported to
modules until that version.

For RHEL 8 kernels, it got exported in 8.3 kernel and newer.

For new kernels later than kernel 5.4, or kernels that applied commit
c3aab9a0bd91 ("mm/filemap.c: don't initiate writeback if mapping has
no dirty pages"), nilfs2 needs to call inode_attach_wb() to avoid a
null pointer dereference BUG, but does not necessarily need it for
earlier versions of kernels.

So, this changes the implementation not to use it in the RHEL 8.2
kernel or earlier.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>